+++ src/__tests__/services/auroraApiClients/responses/imageRepositoryClient/getTags.json
