+++ src/screens/AffiliationViews/Filter/Filter.tsx
@@ -1,6 +1,6 @@
 import Grid from 'aurora-frontend-react-komponenter/Grid';
 import * as React from 'react';
-import { IApplicationDeployment } from 'services/AuroraApiClient/types';
+import { IApplicationDeployment } from 'services/auroraApiClients';
 import MulitDropdown, { MultiDropdownOptions } from './MultiDropdown';
 
 interface IFilterProps {
