--- src/services/AuroraApiClient/queries/user-affiliations-query.ts
@@ -1,29 +1,0 @@
-import { default as gql } from 'graphql-tag';
-
-export const USER_AFFILIATIONS_QUERY = gql`
-  {
-    currentUser {
-      name
-    }
-    affiliations {
-      edges {
-        node {
-          name
-        }
-      }
-    }
-  }
-`;
-
-export interface IUserAffiliationsQuery {
-  currentUser: {
-    name: string;
-  };
-  affiliations: {
-    edges: Array<{
-      node: {
-        name: string;
-      };
-    }>;
-  };
-}
