+++ src/services/auroraApiClients/applicationDeploymentClient/client.ts
@@ -1,0 +1,95 @@
+import ApolloClient from 'apollo-boost';
+import {
+  APPLICATIONS_QUERY,
+  IApplicationInstance,
+  IApplications,
+  IImageRepository,
+  IPodResource,
+  IUserAffiliationsQuery,
+  USER_AFFILIATIONS_QUERY
+} from './query';
+
+export interface IApplicationDeployment {
+  id: string;
+  affiliation: string;
+  name: string;
+  environment: string;
+  statusCode: string;
+  version: {
+    auroraVersion: string;
+    deployTag: string;
+  };
+  repository: string;
+  pods: IPodResource[];
+}
+
+export interface IUserAndAffiliations {
+  affiliations: string[];
+  user: string;
+}
+
+export interface IApplicationDeploymentClient {
+  findUserAndAffiliations: () => Promise<IUserAndAffiliations>;
+  findAllApplicationDeployments: (
+    affiliations: string[]
+  ) => Promise<IApplicationDeployment[]>;
+}
+
+export class ApplicationDeploymentClient
+  implements IApplicationDeploymentClient {
+  private client: ApolloClient<{}>;
+
+  constructor(client: ApolloClient<{}>) {
+    this.client = client;
+  }
+
+  public async findUserAndAffiliations(): Promise<IUserAndAffiliations> {
+    const result = await this.client.query<IUserAffiliationsQuery>({
+      query: USER_AFFILIATIONS_QUERY
+    });
+
+    return {
+      affiliations: result.data.affiliations.edges.map(edge => edge.node.name),
+      user: result.data.currentUser.name
+    };
+  }
+
+  public async findAllApplicationDeployments(
+    affiliations: string[]
+  ): Promise<IApplicationDeployment[]> {
+    const result = await this.client.query<IApplications>({
+      query: APPLICATIONS_QUERY,
+      variables: {
+        affiliations
+      }
+    });
+
+    return result.data.applications.edges.reduce((acc, { node }) => {
+      const { applicationInstances, imageRepository } = node;
+      const apps = applicationInstances.map(app =>
+        this.toApplicationDeployment(node.name, app, imageRepository)
+      );
+      return [...acc, ...apps];
+    }, []);
+  }
+
+  private toApplicationDeployment(
+    name: string,
+    app: IApplicationInstance,
+    imageRepository?: IImageRepository
+  ): IApplicationDeployment {
+    return {
+      affiliation: app.affiliation.name,
+      environment: app.environment,
+      id: app.environment + '-' + name,
+      name,
+      pods: app.details.podResources,
+      repository: imageRepository ? imageRepository.repository : '',
+      statusCode: app.status.code,
+      version: {
+        auroraVersion: app.version.auroraVersion,
+        deployTag: app.version.deployTag
+      }
+    };
+  }
+}
