+++ src/services/auroraApiClients/index.ts
@@ -1,0 +1,8 @@
+export * from './applicationDeploymentClient/client';
+export * from './imageRepositoryClient/client';
+export * from 'models/TagsPagedGroup';
+
+export interface IPageInfo {
+  endCursor: string;
+  hasNextPage: boolean;
+}
